requirements.txt
python-miio
urllib3
